requirements.txt
click==8.1.3
Flask==2.1.2
flask-oidc==1.4.0
httplib2==0.20.4
importlib-metadata==4.12.0
itsdangerous==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauth2client==4.1.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-keycloak==2.1.1
rsa==4.8
six==1.16.0
Werkzeug==2.1.2
zipp==3.8.0
gunicorn
